feat: изменение типов данных для Amount и Balance на int64

+++ database/models.go
@@ -13,8 +13,8 @@ import (
 // Содержит уникальный адрес и текущий баланс.
 type Wallet struct {
 	gorm.Model
-	Address string  `gorm:"unique;not null"` // Address уникальный адрес кошелька, используемый при идентификации
-	Balance float64 //Balance текущий баланс кошелька
+	Address string `gorm:"unique;not null"` // Address уникальный адрес кошелька, используемый при идентификации
+	Balance int64  //Balance текущий баланс кошелька
 }
 
 // Transaction представляет собой модель транзакции в базе данных
@@ -23,7 +23,7 @@ type Transaction struct {
 	ID          uint      `gorm:"primaryKey" json:"id"`        // идентификатор записи
 	FromAddress string    `json:"from_address"`                // адрес отправителя
 	ToAddress   string    `json:"to_address"`                  // адрес получателя
-	Amount      float64   `json:"amount"`                      // сумма перевода
+	Amount      int64     `json:"amount"`                      // сумма перевода
 	Timestamp   time.Time `json:"timestamp"`                   // время создания транзакции
 	UUID        string    `gorm:"unique;not null" json:"uuid"` // уникальный идентификатор транзакции
 }
